veggie system

+++ database/migrations/2021_10_13_023737_create_items_table.php
@@ -16,8 +16,8 @@ class CreateItemsTable extends Migration
         Schema::create('items', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->float('price');
-            $table->float('inventories');
+            $table->double('price');
+            $table->double('inventories');
             $table->double('total_sales');
             $table->timestamps();
             $table->softDeletes();
